Move key import/export logic to trustmanager from utils
