+++ cmd/notary/keys_nonpkcs11.go
@@ -8,17 +8,16 @@ import (
 	"github.com/docker/notary"
 	store "github.com/docker/notary/storage"
 	"github.com/docker/notary/trustmanager"
-	"github.com/docker/notary/utils"
 )
 
 func getYubiStore(fileKeyStore trustmanager.KeyStore, ret notary.PassRetriever) (trustmanager.KeyStore, error) {
 	return nil, errors.New("Not built with hardware support")
 }
 
-func getImporters(baseDir string, _ notary.PassRetriever) ([]utils.Importer, error) {
+func getImporters(baseDir string, _ notary.PassRetriever) ([]trustmanager.Importer, error) {
 	fileStore, err := store.NewPrivateKeyFileStorage(baseDir, notary.KeyExtension)
 	if err != nil {
 		return nil, err
 	}
-	return []utils.Importer{fileStore}, nil
+	return []trustmanager.Importer{fileStore}, nil
 }
